login and signup
folder for login and sign up
+ login by 'enter'
