+++ Spottyfy.csproj
@@ -203,11 +203,17 @@
     </Compile>
     <Compile Include="getAlbums.cs" />
     <Compile Include="IDB.cs" />
-    <Compile Include="Login.cs">
+    <Compile Include="LoginAndSign\Login.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Login.Designer.cs">
+    <Compile Include="LoginAndSign\Login.Designer.cs">
       <DependentUpon>Login.cs</DependentUpon>
+    </Compile>
+    <Compile Include="MainMenu.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="MainMenu.Designer.cs">
+      <DependentUpon>MainMenu.cs</DependentUpon>
     </Compile>
     <Compile Include="MongoDB.cs" />
     <Compile Include="MySqlDB.cs" />
@@ -220,10 +226,10 @@
     <Compile Include="PostgreSqlDB.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SignUp.cs">
+    <Compile Include="LoginAndSign\SignUp.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="SignUp.Designer.cs" />
+    <Compile Include="LoginAndSign\SignUp.Designer.cs" />
     <Compile Include="SongData.cs" />
     <Compile Include="SongView.cs">
       <SubType>Form</SubType>
@@ -238,7 +244,7 @@
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Login.resx">
+    <EmbeddedResource Include="LoginAndSign\Login.resx">
       <DependentUpon>Login.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="NavBar.resx">
@@ -253,7 +259,7 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="SignUp.resx" />
+    <EmbeddedResource Include="LoginAndSign\SignUp.resx" />
     <EmbeddedResource Include="SongView.resx">
       <DependentUpon>SongView.cs</DependentUpon>
     </EmbeddedResource>
